refactor go_to and send_to out of WorkspacesExt
Until now, when the user chose a title from the menu, we performed a
check to see if a workspace existed with that title.  If a match is
found, the target became the name of the matching workspace.  If no
match is found, the target became the chosen title.

The new behavior concerns the case when no match is found: now, we count
the number of existing workspaces and *create* a name from that number
and the chosen title, giving it the lowest unused number and putting it
at the end of the list of workspaces.

This means we no longer need to run any checks on the Workspaces object
inside the go_to() and send_to() functions before returning the
generated command.  Since the only benefit of these functions being
defined on the WorkspacesExt was easy access to the Workspaces object,
this means we can move them out of that trait.  In retrospect, they
didn't have much business being there in the first place.

This results in some refactoring of the code that actually calls these
two functions inside handle_stream().

Also, this commit consolidates some input checks and removes a possible
panic (at a possible tiny performance cost).